separated config file processing into loadConfig.cpp
